refactor(allocator): extract auth.py and utils/ansi.py from main.py
Foundation for the main.py blueprint split (Task 1 of 12): a top-level
auth module lets later route blueprints apply @auth.login_required at
import time, and utils/ansi.py gives two upcoming blueprints a shared
home for ANSI-stripping instead of duplicating it. registration.py's
list_clients now decorates normally instead of hand-invoking
auth.login_required, since main.auth exists by the time it imports.
